requirements.txt
pandas
python-docx
openpyxl
xlrd>=2.0.1
PyMuPDF
Pillow
numpy
paddleocr
